Updated memory associated with ewald correction
Before there were several define statements setting the size of the
ewald table along with a define for the arrays associated with the
correction. This mean that the Ewald table was always on the stack.

The are issues with this:
- newer compilers are likely to complain about multiplications for the
  extent of the ewald arrays being a float and hence not compiling
- stack size could be quite large for large tables
- complaints about compilers not knowing the size of an array at compile time.
- also using 3d array assuming compiler will make a nice contiguous
  block (should happen with optimisation) but this is unnecessary. Just
  make a 1 D array addressed appropriately.

So to fix this issue, I have made static const variables and dynamically
allocated 1 D arrays on the heap. This means indexing has changed (need
to check no errors introduced) and need to clean-up after a run (need to
check this is called at the correct point).

This update means could can compile on Setonix (gcc-13) and other
machines where I have tested gcc-14, gcc-15.

Should test newer clang as well.

Also need to check force calculations are still correct.